requirements.txt
aiohttp==3.8.6
aiosqlite==0.19.0
asyncio==3.4.3
colorgram.py==1.2.0
easy_pil==0.2.0
emoji==2.2.0
openai==0.27.2
Pillow==10.0.1
requests==2.31.0
toml==0.10.2
jishaku==2.5.1
psutil==5.9.5
discord.py==2.3.2
pynvml==11.5.0
asyncpg==0.27.0
pet-pet-gif==1.0.2
python-dateutil==2.8.2
